+++ packages/quanttide-docs-toolkit/packages/dart/lib/quanttide_docs.dart
@@ -1,1 +1,2 @@
 export 'src/document_cubit.dart';
+export 'src/document_repository.dart';