+++ packages/quanttide-docs-toolkit/packages/dart/lib/src/document_repository.dart
@@ -1,0 +1,9 @@
+import 'dart:async';
+
+abstract class DocumentRepository {
+  Future<String> read();
+  Future<String> readFrom(String path);
+  Future<void> write(String content);
+  Stream<String> watch();
+  Future<void> dispose();
+}